Add some qemu support in debug-browser.sh and add qemu-jsc.sh launcher

+++ debug-browser.sh
@@ -7,7 +7,7 @@ if [ x"$WPET_CONFIG_PARSED" != x"true" ]; then
 fi
 
 # These few lines reorder $@ so that options come first.
-TEMP=`getopt -o ja:tcl:h --long jsc,attach:,tui,continue,launcher:,help -- "$@"`
+TEMP=`getopt -o ja:tcl:Qh --long jsc,attach:,tui,continue,launcher:,qemu,help -- "$@"`
 if [ $? != 0 ] ; then echo "Problem parsing options" >&2 ; exit 1 ; fi
 eval set -- "$TEMP"
 
@@ -24,7 +24,8 @@ while true ; do
         -t|--tui) GDB_OPTS+=" --tui"; shift;;
         -c|--continue) export WPET_DEBUG_CONTINUE_ON_ATTACH=1; shift;;
         -l|--launcher) WPET_DEBUG_PROGRAM=$2; WPET_DEBUG_ATTACH=; shift 2;;
-        -h|--help) echo "syntax: $0 [-j|--jsc] [-a|--attach <prog_name>] [-h|--help]" ; exit 0;; # FIXME
+        -Q|--qemu) WPET_USE_QEMU=1; WPET_DEBUG_ATTACH=; shift;;
+        -h|--help) echo "syntax: $0 [-j|--jsc] [-a|--attach <prog_name>] [-t|--tui] [-c|--continue] [l|--launcher <launcher>] [-Q|--qemu] [-h|--help]" ; exit 0;; # FIXME
         --) shift ; break ;;
     esac
 done
@@ -32,22 +33,31 @@ done
 export WPET_DEBUG_PROGRAM
 export WPET_DEBUG_PROGRAM_PATH=/usr/bin/$WPET_DEBUG_PROGRAM
 
-echo ">>> starting program on remote host"
+if [ x"$WPET_USE_QEMU" == x"" ]; then
+    echo ">>> starting program on remote host"
+fi
 
 if [ x"$WPET_DEBUG_ATTACH" != x"" ]; then
     ssh -p $WPET_REMOTE_SSH_PORT "$WPET_REMOTE_SSH_USER@$WPET_REMOTE_HOST" "run_wpe(){ $WPET_DEBUG_PROGRAM_PATH $@ "' & sleep 5; LD_BIND_NOW=1 gdbserver --attach 0.0.0.0:2345 `pidof '$WPET_DEBUG_ATTACH'` ; }; run_wpe' > run-wpe.log 2>&1 &
 
+elif [ x"$WPET_USE_QEMU" != x"" ]; then
+    echo " >>> starting jsc with ${WPET_QEMU}..."
+    echo $WPET_QEMU -L ${WPET_OUTPUT}/staging -g 2345 -seed 0 ${WPET_OUTPUT}/staging/${WPET_DEBUG_PROGRAM_PATH} $@
+    $WPET_QEMU -L ${WPET_OUTPUT}/staging -g 2345 ${WPET_OUTPUT}/staging/${WPET_DEBUG_PROGRAM_PATH} $@ > run-wpe.log 2>&1 &
+    export WPET_REMOTE_HOST=127.0.0.1
 else
     ssh -p $WPET_REMOTE_SSH_PORT "$WPET_REMOTE_SSH_USER@$WPET_REMOTE_HOST" "run_wpe(){ gdbserver 0.0.0.0:2345 $WPET_DEBUG_PROGRAM_PATH $@ ; }; run_wpe" > run-wpe.log 2>&1 &
 
 
 fi
 
-echo ">>> waiting for gdbserver..."
-tail -f run-wpe.log | while read LOGLINE
-do
-    [[ "${LOGLINE}" == *"Listening on port"* ]] && pkill -P $$ tail
-done
+if [ x"$WPET_USE_QEMU" = x"" ]; then
+    echo ">>> waiting for gdbserver..."
+    tail -f run-wpe.log | while read LOGLINE
+        do
+            [[ "${LOGLINE}" == *"Listening on port"* ]] && pkill -P $$ tail
+        done
+fi
 
 
 echo ">>> starting gdb"
